interpreter: handle ExternalProgram commands in run_command
run_command() can take a list of parameters that will describe the
command being run, and an ExternalProgram object can be provided here.

When setting up the ExternalProgram from a cross-file as a list, for
instance:

        [binaries]
        rustc = ['rustc', '--target=arm-linux-androideabi']

The ExternalProgram.get_path() method will be used to prepare the
run_command arguments, but it would return '--target=arm-androideabi'
instead of the proper argument.

This commit ensures every part of the command is actually used in
this case.